refactor: 프로젝트 폴더 및 컴포넌트 구조 개편 #8
