+++ src/App.jsx
@@ -4,7 +4,8 @@ import * as React from 'react';
 
 import { Route, Routes } from 'react-router-dom';
 
-import Footer from './components/Footer/Footer';
+import Footer from './components/layout/Footer/Footer';
+import Header from './components/layout/Header/Header';
 import Home from './pages/Home';
 import { Reset } from 'styled-reset';
 import SearchResults from './pages/SearchResults';
@@ -13,10 +14,14 @@ function App() {
     return (
         <React.Fragment>
             <Reset />
+
+            <Header />
+
             <Routes>
                 <Route path="/" element={<Home />} />
                 <Route path="/search" element={<SearchResults />} />
             </Routes>
+
             <Footer />
         </React.Fragment>
     );
