+++ src/components/BannerSection/BannerSection.jsx
@@ -1,19 +1,19 @@
-import './HeaderSlider.css';
+import './BannerSection.css';
 
-import BannerSkeleton from '../../Skeleton/BannerSkeleton';
+import BannerSkeleton from '../Skeleton/BannerSkeleton';
 import BannerSwiper from './BannerSwiper';
 import TrailerModal from './TrailerModal';
-import useHeaderMovies from '../../../hooks/useHeaderMovies';
+import useHeaderMovies from '../../hooks/useHeaderMovies';
 import { useRef } from 'react';
-import useTrailer from '../../../hooks/useTrailer';
+import useTrailer from '../../hooks/useTrailer';
 
-const HeaderSlider = () => {
+const BannerSection = () => {
     const swiperRef = useRef(null);
     const { isTrailerOpen, trailerUrl, openTrailer, closeTrailer } = useTrailer(swiperRef);
     const { movies, loading } = useHeaderMovies();
 
     return (
-        <div className="header-slider">
+        <section className="banner-section">
             {loading ? (
                 <BannerSkeleton />
             ) : (
@@ -21,8 +21,8 @@ const HeaderSlider = () => {
             )}
 
             <TrailerModal isTrailerOpen={isTrailerOpen} trailerUrl={trailerUrl} onClose={closeTrailer} />
-        </div>
+        </section>
     );
 };
 
-export default HeaderSlider;
+export default BannerSection;
