+++ src/components/BannerSection/TrailerModal.jsx
