+++ src/components/ContentSection/ContentSection.css
@@ -1,7 +1,3 @@
-.main {
-    font-family: var(--font-body);
-}
-
 /* ContentList */
 .content-list {
     margin: 1rem;
@@ -48,7 +44,8 @@
     background: var(--primary-gradient);
 }
 
-.content-swiper {
+/* ContentSwiper */
+.content-list .content-swiper {
     position: relative;
     padding: 1.5rem;
 }
