--- src/components/Header/Header.css
@@ -1,5 +1,0 @@
-/* Header */
-.header {
-    position: relative;
-    font-family: var(--font-title);
-}
