+++ src/components/layout/Header/Header.css
@@ -1,7 +1,8 @@
-/* HeaderBar, NavMenu, SearchBox 컴포넌트 스타일 */
+/* Header */
+.header {
+    position: relative;
+    font-family: var(--font-title);
 
-/* HeaderBar */
-.header-bar {
     display: flex;
     justify-content: space-between;
     align-items: center;
